Fix system connections on machine removal
When a machine is removed, it removes the current
system connection and will select a new one from
the top of the alphabetical list. However, when
doing this there is no guarantee that, if the new
system connection is associated with a machine, the
new system connection will have the right rootfulness.
This will ensure that the new connection, if associated
with a machine, will have proper rootfulness.

Fixes: #22577

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>